requirements.txt
google-genai==1.21.1
pydantic==2.10.6
pydantic-core==2.27.2
gradio==5.20.1
pillow==11.1.0
requests==2.32.3
